Convolution Filteringprep#4: separate files for convolution
