+++ mchf-eclipse/drivers/audio/audio_convolution.h
@@ -1,0 +1,23 @@
+/*  -*-  mode: c; tab-width: 4; indent-tabs-mode: t; c-basic-offset: 4; coding: utf-8  -*-  */
+/************************************************************************************
+ **                                                                                **
+ **                                        UHSDR                                   **
+ **               a powerful firmware for STM32 based SDR transceivers             **
+ **                                                                                **
+ **--------------------------------------------------------------------------------**
+ **                                                                                **
+ **  Description:   Code for fast convolution filtering                            **
+ **  Licence:		GNU GPLv3                                                      **
+ ************************************************************************************/
+
+/* Define to prevent recursive inclusion -------------------------------------*/
+#ifndef __AUDIO_CONVOLUTION_H
+#define __AUDIO_CONVOLUTION_H
+
+#include "arm_math.h"
+//#include "softdds.h"
+#include "uhsdr_hw_i2s.h"
+#include "uhsdr_board.h"
+
+
+#endif
