+++ mchf-eclipse/hardware/uhsdr_board.h
@@ -31,6 +31,12 @@
  *
  */
 
+
+
+// Fast convolution filtering
+// experimental at the moment DD4WH, 2018_08_18
+//#define USE_CONVOLUTION
+
 // old LMS noise reduction
 // will probably never used any more
 //#define OBSOLETE_NR
